+++ src/components/select/filterable-select/filterable-select.component.js
@@ -38,6 +38,7 @@ const FilterableSelect = React.forwardRef(
       onListScrollBottom,
       tableHeader,
       multiColumn,
+      tooltipPosition,
       ...textboxProps
     },
     inputRef
@@ -294,6 +295,7 @@ const FilterableSelect = React.forwardRef(
       if (
         isTextboxActive &&
         textboxRef &&
+        filterText.length &&
         textValue.length > filterText.length &&
         textStartsWithFilter
       ) {
@@ -452,6 +454,7 @@ const FilterableSelect = React.forwardRef(
         onKeyDown: handleTextboxKeydown,
         onChange: handleTextboxChange,
         onMouseDown: handleTextboxMouseDown,
+        tooltipPosition,
         ...textboxProps,
       };
     }
@@ -546,6 +549,8 @@ FilterableSelect.propTypes = {
   isLoading: PropTypes.bool,
   /** A callback that is triggered when a user scrolls to the bottom of the list */
   onListScrollBottom: PropTypes.func,
+  /** Overrides the default tooltip position */
+  tooltipPosition: PropTypes.oneOf(["top", "bottom", "left", "right"]),
 };
 
 export default FilterableSelect;